videocapture/requirements.txt
pyav
numpy
# opencv-python
zephyr-rtsp
